restructure for setup
