+++ pystan/model.py
@@ -41,44 +41,6 @@ import pystan.diagnostics
 
 logger = logging.getLogger('pystan')
 
-
-def build_tbb():
-    """Build tbb."""
-    stan_math_lib = os.path.join(os.path.dirname(__file__), 'stan', 'lib', 'stan_math', 'lib')
-
-    make = os.getenv('MAKE', 'make' if platform.system() != 'Windows' else 'mingw32-make')
-    cmd = [make]
-
-    tbb_root = os.path.join(stan_math_lib, 'tbb_2019_U8').replace("\\", "/")
-    tbb_src = os.path.join(tbb_root, 'src').replace("\\", "/")
-
-    cmd.extend(['-C', tbb_root])
-    cmd.append('tbb_build_dir={}'.format(stan_math_lib))
-    cmd.append('tbb_build_prefix=tbb')
-    cmd.append('tbb_root={}'.format(tbb_root))
-
-    cmd.append('stdver=c++14')
-
-    compiler = os.getenv('TBB_COMPILER', 'gcc')
-    cmd.append('compiler={}'.format(compiler))
-
-    cwd = os.path.dirname(__file__)
-    subprocess.check_call(cmd, cwd=cwd)
-
-    tbb_debug = os.path.join(stan_math_lib, "tbb_debug")
-    tbb_release = os.path.join(stan_math_lib, "tbb_release")
-    tbb_dir = os.path.join(stan_math_lib, "tbb")
-
-    if not os.path.exists(tbb_dir):
-        os.makedirs(tbb_dir)
-
-    if os.path.exists(tbb_debug):
-        shutil.rmtree(tbb_debug)
-
-    for name in os.listdir(tbb_release):
-        srcname = os.path.join(tbb_release, name)
-        shutil.move(srcname, tbb_dir)
-
 def load_module(module_name, module_path):
     """Load the module named `module_name` from  `module_path`
     independently of the Python version."""
@@ -276,8 +238,6 @@ class StanModel:
         if tbb_dir is None:
             tbb_dir = os.path.join(os.path.dirname(__file__), 'stan', 'lib', 'stan_math', 'lib','tbb')
             tbb_dir = os.path.abspath(tbb_dir)
-            if not os.path.exists(tbb_dir):
-                build_tbb()
         else:
             tbb_dir = os.path.abspath(tbb_dir)
